Created separate Input directory
