+++ Source/Core/src/Engine.cpp
@@ -9,8 +9,8 @@
 #include "Core/EngineApis.h"
 #include "Core/MemoryModule.h"
 #include "Core/WindowManager.h"
-#include "Core/MessageDispather.h"
-#include "Core/InputManager.h"
+#include "Core/Input/MessageDispather.h"
+#include "Core/Input/InputManager.h"
 #include "Core/Resources/ResourceManager.h"
 #include "Core/Resources/Mesh/MeshManager.h"
 #include "Core/Resources/Shader/ShaderManager.h"
